Fix weird service bug
Sometimes services can be undefined, but trying to get them doesnt error
(in this case, the service it tried to get was "Network")
If this is the case, GetService will just return nil instead of erroring.
The change checks now to see if service actually exists, as opposed to
if it doesnt error.